[level 0] Title: 두 수의 곱, Time: 0.01 ms, Memory: 4.21 MB -BaekjoonHub

+++ 프로그래머스/0/120804. 두 수의 곱/두 수의 곱.cpp
@@ -4,6 +4,6 @@
 using namespace std;
 
 int solution(int num1, int num2) {
-    int r=num1*num2;
+    int r=(int)num1*num2;
     return r;
 }
